+++ src/unicode-name.d.ts
@@ -1,0 +1,3 @@
+declare module 'unicode-name' {
+    export function unicodeReadableName(character: string | number): string | undefined;
+}